translate(skip): rename SKIP to skip
